warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@4557 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Skype/src/skype_dialogs.cpp
@@ -650,7 +650,7 @@ INT_PTR CALLBACK CSkypeProto::InviteToChatProc(HWND hwndDlg, UINT msg, WPARAM wP
 					if (chatID)
 					{
 						g_skype->GetConversationByIdentity(::mir_utf8encodeW(chatID), conversation);
-						for (uint i = 0; i < invitedContacts.getCount(); i++)
+						for (uint i = 0; i < (uint)invitedContacts.getCount(); i++)
 							needToAdd.append((char *)mir_ptr<char>(::mir_u2a(invitedContacts[i])));
 						conversation->AddConsumers(needToAdd);
 					}
